backend/requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.9

# Pydantic for validation
pydantic>=2.4.0
pydantic-settings>=2.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# AI Integration
anthropic>=0.5.0

# Redis for caching (optional but recommended)
redis>=5.0.0

# Task queue (optional for async operations)
celery>=5.3.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Development tools
black>=23.10.0
flake8>=6.1.0
mypy>=1.6.0

# CORS
python-dotenv>=1.0.0
